Fix bug with a1176080
If you pass a symbol here, it will be interpreted as the name of a
face to use. This works correctly if you use the symbol highlight,
which is the name of a face, or if you define a face called
browse-kill-ring-inserted-item-face. However, a new user will not have
such a face defined (since we don't define it) and will notice
that (even if she sets browse-kill-ring-highlight-current-entry) the
current entry won't be highlight. A similar bug exists for
browse-kill-ring-highlight-inserted-item.