Fix bug with highlighting inserted item
791330f7 introduced a bug by removing a redefinition of the variable
pt. pt as passed in is the value of point in the Kill Ring buffer; it
bears no relationship whatsoever to the value we want, which is the
value of point in the buffer we're inserting into.

Make this a little clearer by using a new name for the variable. This
requires a reindent.