Perform the quit action earlier.
Previously the `*Kill-Ring*' buffer is only closed after the item has
been inserted and the highlighting has finished (if highlighting of the
inserted item is turned on), this takes 0.5 seconds, and gives a user
experience that appears unresponsive.

After this patch when the user asks for insert and exit, we first
establish what will be inserted, then we close the `*Kill Ring*' buffer,
and then we insert and highlight the item.  This gives a much more
responsive user experience.