Delete items that are edited to the empty string.
When using the edit feature of browse-kill-ring, if an item is edited to
the empty string then it is deleted from the `kill-ring'.  Care is taken
to ensure that the correct entry is reselected after the delete in order
to give the same user experience as if the user had just chosen to
delete the item from the kill-ring with browse-kill-ring-delete.