Unify common code after `browse-kill-ring-edit'.
This patch unifies common code from `browse-kill-ring-finish' and
`browse-kill-ring-abort', and adds protection so that we only kill the
edit buffer if it's in the correct mode (so we can be reasonably sure
that it really is the edit buffer).

I've made the `browse-kill-ring-edit-target' variable track when the
user is editing an entry in the `kill-ring', resetting it's value to nil
after the edit is complete.  The variable can then now be used to tell
if calling `browse-kill-ring-edit-finish' is correct or not.