Fix the insert and move functions.
Reorder the calls to insert and delete so that the delete is called on a
target that is acquired before the call to insert.
Calling insert, with the quit parameter set to true, will close the
kill-ring buffer, after this calling `browse-kill-ring-delete' is no
longer possible, instead use `browse-kill-ring-do-delete' with the
target we acquired before calling insert.