Default for browse-kill-ring-occur is to remove regexp.
When entering the empty string for the regexp in browse-kill-ring-occur,
the default action is now to remove the active regexp, rather than to
just reapply the regexp that is currently in use.

This seems to make more sense to me, this gives a mechanism to remove
the filtering, if an incorrect regexp is entered, and why would the user
ever want to filter using the exact same regexp as is already in place,
it will never cause the display to change.