Generate the insert functions using a macro.
We have a series of functions to insert, append, and prepend, each of
which can either just insert, insert and move, or insert and delete.

Previously each of these functions was explicitly written out.  After
this commit the insert functions are generated using a macro.  The
benefit of this approach is that we only have once place to edit if the
insert logic needs to change.