Split up `browse-kill-ring-setup`.
The old `browse-kill-ring-setup` function performed two tasks, it
entered `browse-kill-ring` mode, and performed one-time setup that was
required.  It also updated the buffer contents based on the value of the
kill-ring (or whatever ring we're browsing).

This is unfortunate, as entering a mode resets all of the buffer local
variables.  It also means that we perform unneeded work when all we
really want to do is refresh the contents of the *Kill Ring* buffer.

This commit moves the work related to refreshing the kill-ring contents
into `browse-kill-ring-update`, and leaves one time setup work in the
`browse-kill-ring-setup` function.  We now only call the setup function
once, when we want to first enter `browse-kill-ring` mode.  At all other
times we call the update function.

To the end user there should be no visible changes from this commit.