Remove quit parameter from all of the insert commands
With this change the insert commands no longer take a quit parameter.
Each insert command performs the insert, and then updates the state as
though the browse-kill-ring display is going to continue to exist.

After the insert, it is the top level command that now decides if it
should quit or not.