Convert all insert functions to use browse-kill-ring-insert-new
The browse-kill-ring-insert-new can now handle a 'nil post action,
which is used for functions that insert without either moving, or
deleting the item they just inserted.

With that change, all of the insert functions can now be changed to
make use of browse-kill-ring-insert-new.

I've reordered a couple of functions inorder to group all of the
insert type functions together in one block.

All of the browse-kill-ring-*-and-quit functions just call their
non-quit sibling with the quit parameter forced to 't.