Update kill-ring-yank-pointer after delete from kill-ring
Deleting the item currently referenced by kill-ring-yank-pointer will
result in some strange behaviour on the next yank.  For example, if
the kill-ring contains 6 items: A, B, C, D, E, F, and the
kill-ring-yank-pointer points to 'A', then, if 'A' is deleted, the
next yank will insert item 'F'.

After this commit, in the same situation, the item 'B' would be
inserted, which seems like a better choice.

This fixes issue #61.