Fontify any keywords in browse-kill-ring-fontify-region
Issue #59 tried to integrate form-feed-mode with browse-kill-ring.
The form-feed-mode uses font-lock-keywords to modify how the form feed
character is displayed, the issue occurs when
browse-kill-ring-separator is changed to the form feed character.

The setup, is this:

    ;; Use form-feed character as separator.
    (setq browse-kill-ring-separator "\x0C")
    ;; Activate form-feed-mode when browsing the kill-ring.
    (add-hook 'browse-kill-ring-hook 'form-feed-mode)

The expectation is that the form feed character (used in the
separator) will be replaced with a horizontal line.

Currently this doesn't happen, as browse-kill-ring overrides
font-lock-fontify-region-function which means no keyword highlighting
occurs.

The fix is to call the default keyword highlighting function
(font-lock-fontify-keywords-region) from
browse-kill-ring-fontify-region.

In the standard case there will be no additional keywords and the
extra call will result in no change.  However, in the form-feed-mode
case, the additional keywords are now highlighted, and the horizontal
bar does appear.

While investigating this issue, I noticed that we passed an extra,
unneeded nil when setting up font-lock-defaults.  This was harmless,
but I've cleaned it for clarity.

This resolves issues #59.