Merge all of the do-*-insert functions into one
This brings together browse-kill-ring-do-insert,
browse-kill-ring-do-append-insert, and
browse-kill-ring-do-prepend-insert, into one function.

There was already some functionality (relating to updating
kill-ring-yank-pointer) that was only in browse-kill-ring-do-insert
and was missing from the others, this fixes that mistake, and
hopefully should make it harder for similar issues to appear in the
future.

The single new function takes the name browse-kill-ring-do-insert, and
has an extra location parameter that can be either 'here, 'append, or
'prepend, to control where to insert.